requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
python-dateutil==2.9.0
pytz==2024.1
